Sources/Outcompute.Toolkit/Expressions/OnesComplementWireExpression.cs
using Outcompute.Toolkit.Expressions.Visitors;

namespace Outcompute.Toolkit.Expressions;

/// <summary>
/// A <see cref="WireExpression"/> representing the ones complement.
/// </summary>
public sealed record class OnesComplementWireExpression : WireExpression
{
    internal OnesComplementWireExpression(WireExpression expression)
    {
        Guard.IsNotNull(expression);

        Expression = expression;
    }

    /// <summary>
    /// The expression to convert.
    /// </summary>
    public WireExpression Expression { get; }

    /// <summary>
    /// Makes the specified visitor visit the current expression using the correct overload.
    /// </summary>
    protected internal override WireExpression Accept(WireExpressionVisitor visitor) => visitor.VisitOnesComplement(this);
}

public partial record class WireExpression
{
    /// <summary>
    /// Creates a <see cref="WireExpression"/> representing the ones complement.
    /// </summary>
    public static OnesComplementWireExpression OnesComplement(WireExpression expression) => new(expression);
}